@@ -3,14 +3,16 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:fokus/model/ui/gamification/ui_reward.dart';
 import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/app_paths.dart';
+import 'package:fokus/utils/icon_sets.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/buttons/rounded_button.dart';
 import 'package:fokus/widgets/chips/attribute_chip.dart';
 
 class RewardDialog extends StatefulWidget {
 	final UIReward reward;
+	final bool showHeader;
 
-	RewardDialog({@required this.reward});
+	RewardDialog({@required this.reward, this.showHeader});
 
 	@override
 	_RewardDialogState createState() => new _RewardDialogState();
@@ -42,14 +44,15 @@ class _RewardDialogState extends State<RewardDialog> with SingleTickerProviderSt
 					padding: EdgeInsets.symmetric(horizontal: AppBoxProperties.screenEdgePadding),
 					child: Column(
 						mainAxisSize: MainAxisSize.min,
-						children: [				
-							Padding(
-								padding: EdgeInsets.all(20.0).copyWith(bottom: 0), 
-								child: Text(
-									AppLocales.of(context).translate('$_pageKey.claimRewardTitle'),
-									style: Theme.of(context).textTheme.headline6
-								)
-							),
+						children: [		
+							if(widget.showHeader)		
+								Padding(
+									padding: EdgeInsets.all(20.0).copyWith(bottom: 0), 
+									child: Text(
+										AppLocales.of(context).translate('$_pageKey.claimRewardTitle'),
+										style: Theme.of(context).textTheme.headline6
+									)
+								),
 							Stack(
 								alignment: Alignment.center,
 								children: [
@@ -59,7 +62,7 @@ class _RewardDialogState extends State<RewardDialog> with SingleTickerProviderSt
 									),
 									Padding(
 										padding: EdgeInsets.only(top: 10.0),
-										child: SvgPicture.asset(rewardIconSvgPath(widget.reward.icon), height: MediaQuery.of(context).size.width*0.3)
+										child: SvgPicture.asset(AssetType.rewards.getPath(widget.reward.icon), height: MediaQuery.of(context).size.width*0.3)
 									)
 								]
 							),
@@ -96,13 +99,14 @@ class _RewardDialogState extends State<RewardDialog> with SingleTickerProviderSt
 											onPressed: () => Navigator.of(context).pop(),
 											dense: true
 										),
-										RoundedButton(
-											icon: Icons.add,
-											text: AppLocales.of(context).translate('$_pageKey.claimButton'),
-											color: AppColors.childButtonColor,
-											onPressed: () => { /* TODO Claim reward */ },
-											dense: true
-										)
+										if(widget.showHeader)
+											RoundedButton(
+												icon: Icons.add,
+												text: AppLocales.of(context).translate('$_pageKey.claimButton'),
+												color: AppColors.childButtonColor,
+												onPressed: () => { /* TODO Claim reward */ },
+												dense: true
+											)
 									]
 								)
 							)
