@@ -4,19 +4,15 @@ class ChildBadge {
   String comment;
   TimeDate date;
   int icon;
-  int level;
-  int maxLevel;
   String name;
 
-  ChildBadge({this.comment, this.date, this.icon, this.level, this.maxLevel, this.name});
+  ChildBadge({this.comment, this.date, this.icon, this.name});
 
   factory ChildBadge.fromJson(Map<String, dynamic> json) {
     return json != null ? ChildBadge(
       comment: json['comment'],
       date: TimeDate.parseDBDate(json['date']),
       icon: json['icon'],
-      level: json['level'],
-      maxLevel: json['maxLevel'],
       name: json['name'],
     ) : null;
   }
@@ -26,8 +22,6 @@ class ChildBadge {
     data['comment'] = this.comment;
     data['date'] = this.date.toDBDate();
     data['icon'] = this.icon;
-    data['level'] = this.level;
-    data['maxLevel'] = this.maxLevel;
     data['name'] = this.name;
     return data;
   }
