@@ -1,42 +1,22 @@
+import 'package:equatable/equatable.dart';
 import 'package:fokus/model/db/gamification/badge.dart';
 
-enum UIBadgeMaxLevel { one, three, five }
+class UIBadge extends Equatable {
+	final String name;
+	final String description;
+	final int icon;
 
-extension UIBadgeMaxLevelExtension on UIBadgeMaxLevel {
-  int get value {
-    switch (this) {
-      case UIBadgeMaxLevel.one:
-        return 1;
-      case UIBadgeMaxLevel.three:
-        return 3;
-      case UIBadgeMaxLevel.five:
-        return 5;
-      default:
-        return null;
-    }
-  }
-}
+	UIBadge({this.name, this.description, this.icon = 0});
+	UIBadge.fromDBModel(Badge badge) : this(name: badge.name, description: badge.description, icon: badge.icon);
 
-class UIBadge {
-	String name;
-	String description;
-	UIBadgeMaxLevel maxLevel;
-	int icon;
-	DateTime date;
+	UIBadge copyWith({String name, String description, int icon}) {
+		return UIBadge(
+			name: name ?? this.name,
+			description: description ?? this.description,
+			icon: icon ?? this.icon
+		);
+	}
 
-	UIBadge({
-		this.name,
-		this.description,
-		this.maxLevel = UIBadgeMaxLevel.one,
-		this.icon = 0,
-		this.date
-	});
-	UIBadge.fromDBModel(Badge badge) : this(
-		name: badge.name,
-		description:
-		badge.description,
-		icon: badge.icon,
-		maxLevel: badge.maxLevel == 5 ? UIBadgeMaxLevel.five : (badge.maxLevel == 3 ? UIBadgeMaxLevel.three : UIBadgeMaxLevel.one)
-	);
-
+  @override
+  List<Object> get props => [name, description, icon];
 }
