@@ -1,19 +1,22 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fokus/logic/caregiver_awards_cubit.dart';
 import 'package:fokus/model/ui/app_page.dart';
 import 'package:fokus/model/ui/gamification/ui_badge.dart';
 import 'package:fokus/model/ui/ui_button.dart';
 import 'package:fokus/services/app_locales.dart';
+import 'package:fokus/utils/dialog_utils.dart';
 import 'package:fokus/utils/icon_sets.dart';
+import 'package:fokus/utils/snackbar_utils.dart';
 
 import 'package:fokus/widgets/app_header.dart';
 import 'package:fokus/widgets/app_navigation_bar.dart';
 import 'package:fokus/widgets/cards/item_card.dart';
 import 'package:fokus/widgets/chips/attribute_chip.dart';
+import 'package:fokus/widgets/dialogs/general_dialog.dart';
 import 'package:fokus/widgets/loadable_bloc_builder.dart';
 import 'package:fokus/widgets/segment.dart';
+import 'package:mongo_dart/mongo_dart.dart' as Mongo;
 
 class CaregiverAwardsPage extends StatefulWidget {
 	@override
@@ -31,9 +34,9 @@ class _CaregiverAwardsPageState extends State<CaregiverAwardsPage> {
 				children: <Widget>[
 					AppHeader.normal(title: '$_pageKey.header.title', text: '$_pageKey.header.pageHint', headerActionButtons: [
 						HeaderActionButton.normal(Icons.add, '$_pageKey.header.addReward', 
-						() => Navigator.of(context).pushNamed(AppPage.caregiverRewardForm.name, arguments: AppFormType.create)),
+						() => Navigator.of(context).pushNamed(AppPage.caregiverRewardForm.name)),
 						HeaderActionButton.normal(Icons.add, '$_pageKey.header.addBadge', 
-						() => Navigator.of(context).pushNamed(AppPage.caregiverBadgeForm.name, arguments: AppFormType.create))
+						() => Navigator.of(context).pushNamed(AppPage.caregiverBadgeForm.name))
 					]),
 					LoadableBlocBuilder<CaregiverAwardsCubit>(
 						builder: (context, state) => AppSegments(segments: _buildPanelSegments(state, context))
@@ -44,7 +47,19 @@ class _CaregiverAwardsPageState extends State<CaregiverAwardsPage> {
     );
 	}
 
-	List<Segment> _buildPanelSegments(CaregiverAwardsLoadSuccess state, context) {
+	void _deleteReward(Mongo.ObjectId id) {
+		context.bloc<CaregiverAwardsCubit>().removeReward(id);
+		Navigator.of(context).pop(); // closing confirm dialog before pushing snackbar
+		showSuccessSnackbar(context, '$_pageKey.content.rewardRemovedText');
+	}
+
+	void _deleteBadge(UIBadge badge) {
+		context.bloc<CaregiverAwardsCubit>().removeBadge(badge);
+		Navigator.of(context).pop(); // closing confirm dialog before pushing snackbar
+		showSuccessSnackbar(context, '$_pageKey.content.badgeRemovedText');
+	}
+
+	List<Segment> _buildPanelSegments(CaregiverAwardsLoadSuccess state, BuildContext context) {
 		return [
 			Segment(
 				title: '$_pageKey.content.addedRewardsTitle',
@@ -59,15 +74,26 @@ class _CaregiverAwardsPageState extends State<CaregiverAwardsPage> {
 				elements: <Widget>[
 					for (var reward in state.rewards)
 						ItemCard(
-							title: reward.name, 
-							subtitle: AppLocales.of(context).translate((reward.limit != null || reward.limit == 0 ) ? 
+							title: reward.name,
+							subtitle: AppLocales.of(context).translate((reward.limit != null || reward.limit == 0 ) ?
 								'$_pageKey.content.limitedReward' : '$_pageKey.content.unlimitedReward', {'REWARD_LIMIT': reward.limit.toString()}),
 							menuItems: [
-								UIButton.ofType(ButtonType.edit, () => {log("edit")}),
-								UIButton.ofType(ButtonType.delete, () => {log("delete")})
+								UIButton.ofType(ButtonType.edit, () => { Navigator.of(context).pushNamed(AppPage.caregiverRewardForm.name, arguments: reward.id) }),
+								UIButton.ofType(ButtonType.delete, () {
+									showBasicDialog(context,
+										GeneralDialog.confirm(
+											title: AppLocales.of(context).translate('$_pageKey.content.removeRewardTitle'),
+											content: AppLocales.of(context).translate('$_pageKey.content.removeRewardText'),
+											confirmColor: Colors.red,
+											confirmText: 'actions.delete',
+											confirmAction: () => _deleteReward(reward.id)
+										)
+									);
+								})
 							],
-							graphicType: GraphicAssetType.rewardsIcons,
+							graphicType: AssetType.rewards,
 							graphic: reward.icon,
+							onTapped: () => showRewardDialog(context, reward, showHeader: false),
 							chips: <Widget>[
 								if(reward.cost != null)
 									AttributeChip.withCurrency(content: reward.cost.quantity.toString(), currencyType: reward.cost.type, tooltip: '$_pageKey.content.pointCost')
@@ -89,12 +115,22 @@ class _CaregiverAwardsPageState extends State<CaregiverAwardsPage> {
 					for (var badge in state.badges)
 						ItemCard(
 							title: badge.name, 
-							subtitle: badge.maxLevel != null ? AppLocales.of(context).translate('$_pageKey.content.${badge.maxLevel.value}LeveledBadge') : '',
+							subtitle: badge.description != null ? badge.description : AppLocales.of(context).translate('$_pageKey.content.noDescriptionSubtitle'),
 							menuItems: [
-								UIButton.ofType(ButtonType.edit, () => {log("edit")}),
-								UIButton.ofType(ButtonType.delete, () => {log("delete")})
+								UIButton.ofType(ButtonType.delete, () {
+									showBasicDialog(context,
+										GeneralDialog.confirm(
+											title: AppLocales.of(context).translate('$_pageKey.content.removeBadgeTitle'),
+											content: AppLocales.of(context).translate('$_pageKey.content.removeBadgeText'),
+											confirmColor: Colors.red,
+											confirmText: 'actions.delete',
+											confirmAction: () => _deleteBadge(badge)
+										)
+									);
+								})
 							],
-							graphicType: GraphicAssetType.badgeIcons,
+							onTapped: () => showBadgeDialog(context, badge, showHeader: false),
+							graphicType: AssetType.badges,
 							graphic: badge.icon,
 							graphicHeight: 44.0
 						)
