@@ -8,8 +8,9 @@ class AppColors {
   static final Color darkTextColor = Colors.black87;
   static final Color mainBackgroundColor = Color.fromARGB(255, 30, 121, 233);
 	static final Color formColor = Colors.teal;
-	static final Color negativeColor = Colors.red;
-	static final Color positiveColor = Colors.lightGreen;
+	static final Color failColor = Colors.red[400];
+	static final Color successColor = Colors.green[400];
+	static final Color infoColor = Colors.blue[400];
 
   // Caregiver colors
   static final Color caregiverBackgroundColor = Color.fromARGB(255, 30, 121, 233);
@@ -23,6 +24,11 @@ class AppColors {
 	static final Color childTaskFiller = Colors.white;
 	static final Color childBreakColor = Colors.blueGrey;
 	static final Color childTaskColor = Colors.lightBlue;
+
+	static final Color negativeColor = Colors.red;
+	static final Color positiveColor = Colors.lightGreen;
+
+  static final Color notificationAccentColor = Color(0xfffdbf00);
 
   // Currency colors
 	static final Map<CurrencyType, Color> currencyColor = {
