@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:fokus/model/currency_type.dart';
 import 'package:fokus/model/ui/ui_button.dart';
 import 'package:fokus/model/ui/gamification/ui_currency.dart';
 import 'package:fokus/services/app_locales.dart';
 import 'package:fokus/utils/app_paths.dart';
 import 'package:fokus/utils/form_config.dart';
+import 'package:fokus/utils/icon_sets.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/buttons/bottom_sheet_bar_buttons.dart';
 import 'package:smart_select/smart_select.dart';
@@ -103,8 +105,8 @@ class _PointPickerFieldState extends State<PointPickerField> {
 											Padding(
 												padding: EdgeInsets.only(left: 10.0, top: 4.0),
 												child: CircleAvatar(
-													child: SvgPicture.asset(currencySvgPath(state.value.type), width: 28, fit: BoxFit.cover),
-													backgroundColor: AppColors.currencyColor[state.value.type].withAlpha(50)
+													child: SvgPicture.asset(getIconPath(state.value != null ? state.value.type : CurrencyType.diamond), width: 28, fit: BoxFit.cover),
+													backgroundColor: AppColors.currencyColor[state.value != null ? state.value.type : CurrencyType.diamond].withAlpha(50)
 												)
 											),
 											(widget.loading) ?
@@ -152,7 +154,7 @@ class _PointPickerFieldState extends State<PointPickerField> {
 							children: [
 								Padding(
 									padding: EdgeInsets.only(right: 6.0),
-									child: SvgPicture.asset(currencySvgPath(item.value.type), width: 30, fit: BoxFit.cover)
+									child: SvgPicture.asset(getIconPath(item.value.type), width: 30, fit: BoxFit.cover)
 								),
 								Text(item.value.getName(context), style: TextStyle(fontWeight: FontWeight.bold, color: AppColors.currencyColor[item.value.type]))
 							]
@@ -172,4 +174,5 @@ class _PointPickerFieldState extends State<PointPickerField> {
 		);
   }
 
+  String getIconPath(CurrencyType type) => AssetType.currencies.getPath(type.index);
 }
